feat: S3 파일 업로드 요청 및 File 컴포넌트명 FileItem으로 변경
